Sorry about not giving details before
I changed the speed on gettingOpen so that speed will matter more and faster receivers are more dangerous. I raised the prs on passRushing there will be more sacks. I did the same thing on runStopping for the same reason. I changed the tackling so that tackling will rely more on the tackling rating rather than speed or strength.